Restore 'v1beta1' version of firestore test proto.
Needed to allow continued testing of the 'v1beta1' code in Python.

Renames 'test.proto' -> 'test_v1.proto', but leaves a 'test.proto' symlink
to allow scripts which do not need both to continue to work.

Adds a version qualifier to the protobuf package (so that descriptors don't
clash in a process which uses both files).